[MOV] portal : reorganize module directory
