+++ addons/portal/__openerp__.py
@@ -22,12 +22,12 @@ the portal automatically adds it to the portal users, etc).  That feature is
 very handy when used in combination with the module 'share'.
     """,
     'data': [
-        'portal_data.xml',
-        'portal_view.xml',
-        'wizard/portal_wizard_view.xml',
+        'data/portal_data.xml',
+        'views/portal_views.xml',
+        'wizard/portal_wizard_views.xml',
         'security/ir.model.access.csv',
     ],
-    'demo': ['portal_demo.xml'],
+    'demo': ['data/portal_demo.xml'],
     'auto_install': True,
     'installable': True,
 }
