+++ addons/portal/models/res_groups.py
